Add State handling. Fix #11.

+++ percentagechartview/src/main/java/com/ramijemli/percentagechartview/renderer/OffsetEnabledMode.java
@@ -18,7 +18,7 @@ package com.ramijemli.percentagechartview.renderer;
 
 public interface OffsetEnabledMode {
 
-    float getBackgroundOffset();
+    int getBackgroundOffset();
 
     void setBackgroundOffset(int backgroundOffset);
